backend/requirements.txt
# FastAPI 和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Telegram Bot
aiogram==3.3.0

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis
redis==5.0.1
aioredis==2.0.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# 其他工具
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
pillow==10.1.0
